refactor(renderer): extract picking, mesh, and visual modules from renderer_objects
Finish modularizing renderer_objects.cpp (3646 -> 773 lines core). New files: renderer_objects_picking.cpp (color-id FBO + pick readback), renderer_objects_mesh.cpp (mesh load/cache, model-file lookup, extents), renderer_objects_visual.cpp (spheres, text/cube meshes, selection box, model preview). Core retains lifecycle, the main Draw pass, shaders, and IsSkippedModelId/IsVehicleType. Pure code motion; builds clean, runtime-verified on D:/IGI1 level 2 (558 objects, 93 textures, 232 attachments, no errors).
